GPU HAL: BCOO sparse matvec + column-norm scaling + validation
First working end-to-end benchmark on real hal9001 basis at R=7, n=10k:
  Basis construction (R, CPU): 2.9s
  GPU HAL CV+fit (5-fold, 20 lambdas): 175.7s
  Total: 178.6s
  In-sample R²: 0.097 (real signal recovered)

Optimizations:
- jax.experimental.sparse.BCOO for GPU-native sparse matvec (2.6x
  faster than scipy-bounce approach)
- Column-norm scaling of basis matrix before Lasso (prevents lambda_max
  being dominated by large-scale smoothness-1 spline bases; β is
  unscaled back to original basis at the end)
- float32 throughout JAX to maximize GPU throughput

Validation:
- FISTA dense Lasso matches sklearn to 1e-7 relative L2 error
- Sparse path runs end-to-end with convergence
- Real R=7 basis gives non-null fit with expected R² magnitude

Speedup vs hal9001 CPU:
- At n=10k: GPU 3 min vs CPU ~1 min — currently SLOWER at small n
- At n=50k: projected GPU ~15 min vs CPU ~5 min — still slower
- At n=451k (full panel): GPU ~150 min vs CPU ~100 HOURS — GPU wins

The current FISTA-based implementation breaks even around n=100k. For
smaller problems CPU hal9001 is still best; for full-panel regHAL-TMLE
with bootstrap (what we actually need), GPU is the only tractable path.

Future optimization path: replace FISTA with on-GPU coordinate descent
for additional 5-10x speedup. Deferred to a separate project.

+++ gpu_hal/cv.py
@@ -162,7 +162,7 @@ def cv_fista_lasso_sparse(
         y_tr_mean = float(y_tr.mean())
         y_tr_c = y_tr - y_tr_mean
         matvec, rmatvec, L = make_sparse_matvec(X_tr)
-        y_tr_j = jnp.asarray(y_tr_c)
+        y_tr_j = jnp.asarray(y_tr_c, dtype=jnp.float32)
         fits = fista_lasso_path(
             matvec, rmatvec, y_tr_j, p, lambdas, len(y_tr),
             L=L, max_iter=max_iter, tol=tol, verbose=False,
